Add autoscale function for plotting module
Previously, plotting area did not have an autoscale function such that
the axis remain the same no matter the value progression. This commit
implement an recent change a new `gtkchart` fork's feature to help with
this problem.

There is still problems remain: the chart resoulution did not correspond
with database and there is a hanging line when changing frame